Enableing compiler strict type check
Most OpenSSL types that were converted from the OpenSSL headers
were declared as untyped pointers ("type Pointer").
This designation allows them to be compatible with any type of pointer,
including classes and interfaces. Such a type convention can lead to
missed errors, as the compiler does not generate any warnings or errors
for invalid type usage. This behavior is risky, and mistakes of this nature
can be challenging to detect.
This commit replaced the untyped pointer declarationwith a dummy record
declaration and a typed pointer to the defined record type
in TaurusTLSHeaders_types.pas. This pattern ensures the compiler performs
strict type checks and generates errors when incompatible types are used.

P.S.
types below were not updated as they used in OpenSSL routines as is:
  i2d_of_void = type Pointer;
  Pi2d_of_void = ^i2d_of_void;
  d2i_of_void = type Pointer;
  Pd2i_of_void = ^d2i_of_void;

Additional discussion is needed.