Fix: "array of OSSL_PARAM" routine parameters
Multiple functions were declared with parameter type as
"array of OSSL_PARAM" (open array type parameter).
The C calling convention "const <TYPE> ident[]" is different
from the Pascal convention "array of <TYPE>",
and Pascal adds one hidden parameter to the function call,
which shifts other parameters to the right.
The parameter type is changed to a typed pointer, which
is aligned with the C calling convention.